marged 2012/10/08

+++ splash/SplashFTFont.h
@@ -23,7 +23,7 @@
 #ifndef SPLASHFTFONT_H
 #define SPLASHFTFONT_H
 
-#include "poppler/poppler-config.h"
+#include "poppler-config.h"
 
 #if HAVE_FREETYPE_FREETYPE_H || HAVE_FREETYPE_H
 
@@ -70,7 +70,7 @@ private:
   FT_Matrix matrix;
   FT_Matrix textMatrix;
   SplashCoord textScale;
-  double size;
+  int size;
   GBool enableFreeTypeHinting;
   GBool enableSlightHinting;
 };
